Moved Exceptions tests to separate directory
